pyproject.toml
[project]
name = "chatbot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.0",
    "django-extensions>=4.1",
    "django>=6.0.1",
    "python-decouple>=3.8",
    "taskipy>=1.14.1",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "chromadb>=1.4.0",
    "sentence-transformers>=5.2.0",
    "langchain-openai>=1.1.7",
    "pypdf>=6.6.0",
    "langchain-chroma>=1.1.0",
]

[dependency-groups]
dev = [
    "ipython>=9.9.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]

